requirements.txt
beautifulsoup4==4.13.4
langchain==0.3.26
langchain_chroma==0.2.4
langchain_community==0.3.27
langchain_core==0.3.68
langchain_groq==0.3.5
langchain_huggingface==0.3.0
python-dotenv==1.1.1
requests==2.32.4
streamlit==1.45.1
sentence_transformers==5.0.0
pysqlite3-binary==0.5.4
